Support ECDSA keys
I am trying to use all ECDSA keys on my newer instances.  I am using `ssh-keygen -t ecdsa -m PEM` to create them, and to be able to successfully generate a certificate I needed to update the openssl command to read the key.  Happy to modify this if there is a better way.